fix: resolve Python plugin SEGFAULT in sub-interpreter threading
Two root causes fixed:

1. Stack buffer overflow in w_unpack varargs (int/long mismatch):
   - w_Load(): 'int id' received 8-byte long write via 'l' format, corrupting
     adjacent stack variable. Changed to 'long id' and fixed %d→%ld format.
   - cpythoninterpreter.cpp: cast id to (long) in lib_pack() call
   - cpipython.cpp OnCtmToHub: cast AddrPort()/GetServPort() to (long)
   - cpipython.cpp OnSetConfig: cast val_type to (long)

2. Thread-unsafe PyThreadState sharing in w_CallHook:
   - In sub-interpreter mode, w_CallHook swapped to script->state which was
     created by the main thread. When called from background threads (stress
     test, timer callbacks), this caused SEGFAULT in _PyEval_EvalFrameDefault
     because PyThreadState is thread-specific in CPython.
   - Fix: use PyThreadState_New(interp) to create a per-call thread state
     in the sub-interpreter, then PyThreadState_Delete after use.
   - Also fixed duplicate #ifndef PYTHON_SINGLE_INTERPRETER guard.

Validated: 30/30 stress test runs pass (was ~30% before fix).
Full test suite: 17/17 pass.

+++ plugins/python/cpipython.cpp
@@ -1012,7 +1012,7 @@ bool cpiPython::OnUnLoad(long code)
 bool cpiPython::OnCtmToHub(cConnDC *conn, string *ref)
 {
 	if (conn && ref) {
-		w_Targs *args = lib_pack("sslls", conn->mMyNick.c_str(), conn->AddrIP().c_str(), conn->AddrPort(), conn->GetServPort(), ref->c_str());
+		w_Targs *args = lib_pack("sslls", conn->mMyNick.c_str(), conn->AddrIP().c_str(), (long)conn->AddrPort(), (long)conn->GetServPort(), ref->c_str());
 		return CallAll(W_OnCtmToHub, args, conn);
 	}
 
@@ -1176,7 +1176,7 @@ bool cpiPython::OnNewBan(cUser *user, cBan *ban) // todo: is not called
 bool cpiPython::OnSetConfig(cUser *user, string *conf, string *var, string *val_new, string *val_old, int val_type)
 {
 	if (user && conf && var && val_new && val_old) {
-		w_Targs *args = lib_pack("sssssl", user->mNick.c_str(), conf->c_str(), var->c_str(), val_new->c_str(), val_old->c_str(), val_type);
+		w_Targs *args = lib_pack("sssssl", user->mNick.c_str(), conf->c_str(), var->c_str(), val_new->c_str(), val_old->c_str(), (long)val_type);
 		return CallAll(W_OnSetConfig, args, user->mxConn);
 	}
